CSP-501 Removed api from route path
