+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application.UnitTests/InnerApi/Regions/GetRegionsQueryRequestTests.cs
@@ -1,0 +1,14 @@
+﻿using FluentAssertions;
+using SFA.DAS.ApprenticeAan.Application.InnerApi.Regions;
+
+namespace SFA.DAS.ApprenticeAan.Application.UnitTests.InnerApi.Regions;
+
+public class GetRegionsQueryRequestTests
+{
+    [Test]
+    public void CheckRequestUrl()
+    {
+        var request = new GetRegionsQueryRequest();
+        request.GetUrl.Should().Be("/regions");
+    }
+}
