+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application/Common/Constants.cs
@@ -1,12 +1,10 @@
-﻿namespace SFA.DAS.ApprenticeAan.Application.Common
+﻿namespace SFA.DAS.ApprenticeAan.Application.Common;
+
+public static class Constants
 {
-    public static class Constants
+    public static class AanHubApiUrls
     {
-        public static class AanHubApiUrls
-        {
-            private const string ApiPrefix = "/api";
-            public const string GetRegionsUrl = $"{ApiPrefix}/regions";
-            public const string GetProfilesUrl = $"{ApiPrefix}/profiles/";
-        }
+        public const string GetRegionsUrl = $"/regions";
+        public const string GetProfilesUrl = $"/profiles/";
     }
 }
