+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application/InnerApi/Profiles/GetProfilesByUserTypeQueryRequest.cs
@@ -1,0 +1,11 @@
+﻿using SFA.DAS.ApprenticeAan.Application.Common;
+using SFA.DAS.SharedOuterApi.Interfaces;
+
+namespace SFA.DAS.ApprenticeAan.Application.InnerApi.Profiles;
+
+public class GetProfilesByUserTypeQueryRequest : IGetApiRequest
+{
+    public string GetUrl => Constants.AanHubApiUrls.GetProfilesUrl + UserType;
+    private string UserType { get; }
+    public GetProfilesByUserTypeQueryRequest(string userType) => UserType = userType;
+}
