+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application/InnerApi/Regions/GetRegionsQueryRequest.cs
@@ -1,0 +1,9 @@
+﻿using SFA.DAS.ApprenticeAan.Application.Common;
+using SFA.DAS.SharedOuterApi.Interfaces;
+
+namespace SFA.DAS.ApprenticeAan.Application.InnerApi.Regions;
+
+public class GetRegionsQueryRequest : IGetApiRequest
+{
+    public string GetUrl => Constants.AanHubApiUrls.GetRegionsUrl;
+}
