+++ src/ApprenticeAan/SFA.DAS.ApprenticeAan.Application/Profiles/Queries/GetProfilesByUserType/GetProfilesByUserTypeQueryHandler.cs
@@ -1,20 +1,19 @@
 ﻿using MediatR;
 using SFA.DAS.ApprenticeAan.Api.Configuration;
-using SFA.DAS.ApprenticeAan.Application.InnerApi.Profiles.Requests;
+using SFA.DAS.ApprenticeAan.Application.InnerApi.Profiles;
 using SFA.DAS.ApprenticeAan.Application.Services;
 
-namespace SFA.DAS.ApprenticeAan.Application.Profiles.Queries.GetProfilesByUserType
+namespace SFA.DAS.ApprenticeAan.Application.Profiles.Queries.GetProfilesByUserType;
+
+public class GetProfilesByUserTypeQueryHandler : IRequestHandler<GetProfilesByUserTypeQuery, GetProfilesByUserTypeQueryResult?>
 {
-    public class GetProfilesByUserTypeQueryHandler : IRequestHandler<GetProfilesByUserTypeQuery, GetProfilesByUserTypeQueryResult?>
+    private readonly IAanHubApiClient<AanHubApiConfiguration> _apiClient;
+    public GetProfilesByUserTypeQueryHandler(IAanHubApiClient<AanHubApiConfiguration> apiClient)
     {
-        private readonly IAanHubApiClient<AanHubApiConfiguration> _apiClient;
-        public GetProfilesByUserTypeQueryHandler(IAanHubApiClient<AanHubApiConfiguration> apiClient)
-        {
-            _apiClient = apiClient;
-        }
-        public async Task<GetProfilesByUserTypeQueryResult?> Handle(GetProfilesByUserTypeQuery request, CancellationToken cancellationToken)
-        {
-            return await _apiClient.Get<GetProfilesByUserTypeQueryResult>(new GetProfilesByUserTypeQueryRequest(request.UserType));
-        }
+        _apiClient = apiClient;
+    }
+    public async Task<GetProfilesByUserTypeQueryResult?> Handle(GetProfilesByUserTypeQuery request, CancellationToken cancellationToken)
+    {
+        return await _apiClient.Get<GetProfilesByUserTypeQueryResult>(new GetProfilesByUserTypeQueryRequest(request.UserType));
     }
 }
